Revert "IC4C-214: Changed RCI's get_string() arguments type to avoid casting."
This reverts commit 1418946466e762cc89c9cfe3cfe10b76da3a51d6.

+++ private/rci_binary_support.h
@@ -286,7 +286,7 @@ typedef struct
         } element;
 
         connector_element_value_t value;
-        uint32_t string_value_length;
+        size_t string_value_length;
 
         connector_remote_config_t callback_data;
     } shared;
